+++ CHANGELOG.md
@@ -1,12 +1,17 @@
-0.7
+0.8
 ---
 
+0.7
+---
+* `Types.from` now takes `ParameterizedType`s instead of selectors and classes. This allows for other ways of getting
+  generic types.
+* `Types.generic*` methods return `Optional`s instead of `null`.
+* `Type.SupertypeSelector` no longer exists.
 
 0.6
 ---
 * New general method `Types.from`, which helps support superinterfaces as well, and some assorted helper methods.
 * Fixing the Javadoc publishing workflow.
-* Assorted doc fixes.
 
 0.5
 ---